Add DiskEncryption fields to LKENodePools
LKENodePool instances have disc encryption is not configurable and is
enabled by default. Existing LKENodePools will create instances that DO
NOT have disk encryption enabled. As such there is no option to set
DiskEncryption on LKENodePools at creation time but the status of
DiskEncryption is returned from the API when viewing LKENodePools.