Add support for creation of NB with IPv4 Reserved Address (#733)
## 📝 Description
This adds support for creating a NodeBalancer with a reserved IP address. It will be supported by an upcoming release to modify Linode API to accept reserved IPs at the creation of a NodeBalancer. This is needed to support the creation of Service LoadBalancers with reserved IPs in LKE.

## ✔️ How to Test
Customer flags required:
- reserved_ips_beta
- can_reserve_ip
- reserved_ip_nodebalancer

Relevant unit tests can be run using:
```
make TEST_ARGS="-run TestNodeBalancer" test-unit
```
Relevant files:
- test/unit/fixtures/nodebalancer_create_with_ipv4.json
- test/unit/nodebalancer_test.go